move api.Status to common.go. Add Status in expapi/v1/types.go
