+++ pkg/expapi/v1/deep_copy_generated.go
@@ -997,6 +997,52 @@ func deepCopy_v1_ScaleStatus(in ScaleStatus, out *ScaleStatus, c *conversion.Clo
 	return nil
 }
 
+func deepCopy_v1_Status(in Status, out *Status, c *conversion.Cloner) error {
+	if err := deepCopy_v1_TypeMeta(in.TypeMeta, &out.TypeMeta, c); err != nil {
+		return err
+	}
+	if err := deepCopy_v1_ListMeta(in.ListMeta, &out.ListMeta, c); err != nil {
+		return err
+	}
+	out.Status = in.Status
+	out.Message = in.Message
+	out.Reason = in.Reason
+	if in.Details != nil {
+		out.Details = new(StatusDetails)
+		if err := deepCopy_v1_StatusDetails(*in.Details, out.Details, c); err != nil {
+			return err
+		}
+	} else {
+		out.Details = nil
+	}
+	out.Code = in.Code
+	return nil
+}
+
+func deepCopy_v1_StatusCause(in StatusCause, out *StatusCause, c *conversion.Cloner) error {
+	out.Type = in.Type
+	out.Message = in.Message
+	out.Field = in.Field
+	return nil
+}
+
+func deepCopy_v1_StatusDetails(in StatusDetails, out *StatusDetails, c *conversion.Cloner) error {
+	out.Name = in.Name
+	out.Kind = in.Kind
+	if in.Causes != nil {
+		out.Causes = make([]StatusCause, len(in.Causes))
+		for i := range in.Causes {
+			if err := deepCopy_v1_StatusCause(in.Causes[i], &out.Causes[i], c); err != nil {
+				return err
+			}
+		}
+	} else {
+		out.Causes = nil
+	}
+	out.RetryAfterSeconds = in.RetryAfterSeconds
+	return nil
+}
+
 func deepCopy_v1_SubresourceReference(in SubresourceReference, out *SubresourceReference, c *conversion.Cloner) error {
 	out.Kind = in.Kind
 	out.Namespace = in.Namespace
@@ -1122,6 +1168,9 @@ func init() {
 		deepCopy_v1_Scale,
 		deepCopy_v1_ScaleSpec,
 		deepCopy_v1_ScaleStatus,
+		deepCopy_v1_Status,
+		deepCopy_v1_StatusCause,
+		deepCopy_v1_StatusDetails,
 		deepCopy_v1_SubresourceReference,
 		deepCopy_v1_ThirdPartyResource,
 		deepCopy_v1_ThirdPartyResourceList,
